% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCU12_Precip.R
\docType{data}
\name{CCU12_Precip}
\alias{CCU12_Precip}
\title{Precipitation for June-September 2012 recorded in Kolkata}
\format{A data frame with columns
\describe{

\item{Date}{The data in Year-Month-Day format}
\item{Precip}{Precipitation in millimeters}
\item{TMax}{Maximum temperature, in Celcius}
\item{TMin}{Minimum temperature, in Celcius}
}}
\usage{
data(CCU12_Precip)
}
\description{
Precipitation for June-September 2012 recorded in Kolkata
}
\examples{
Precip <-CCU12_Precip$Precip
TMax <-CCU12_Precip$TMax
plot(TMax, Precip)
}
\keyword{datasets}
