% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentityMatrix.R
\name{IdentityMatrix}
\alias{IdentityMatrix}
\title{Obtains the identity matrix of dimension n}
\usage{
IdentityMatrix(n)
}
\arguments{
\item{n}{an integer}
}
\value{
an identity matrix
}
\description{
Obtains the identity matrix of dimension n
}
\examples{
 I.3 = IdentityMatrix(3)
 print(I.3)

}
