% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InnerProduct.R
\name{InnerProduct}
\alias{InnerProduct}
\title{Computes the Euclidean inner product}
\usage{
InnerProduct(a, b, na.rm)
}
\arguments{
\item{a}{a numeric vector}

\item{b}{another numeric vector}

\item{na.rm}{logical}
}
\value{
a real number
}
\description{
Computes the Euclidean inner product
}
\examples{
 x <- c(1, 2, 3)
 y <- c(3, 0, 1)
 InnerProduct(x, y)



}
