% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm.R
\name{Norm}
\alias{Norm}
\title{Computes the Euclidean norm}
\usage{
Norm(a, na.rm)
}
\arguments{
\item{a}{a numeric vector}

\item{na.rm}{logical}
}
\value{
a real number
}
\description{
Computes the Euclidean norm
}
\examples{
 x <- c(1, 2)
 Norm(x)



}
